Clean up BacktraceCleaner
The global configuration object exposes exclusion_patterns,
inclusion_patterns, and full_backtrace for modification at any time
during a run. This means that it would be perfectly reasonable for a
user to turn full_backtrace on and off around specific examples, and
also change the content of each patterns array at any time.

This commit does a number of things:

* clarifies examples
* decouples full_backtrace from exclusion_patterns
* makes it so that setting full_backtrace true or false
  preserves the content of exclusion_patterns and inclusion_patterns.
* simplifies the constructor
  * no args (no code outside examples submitted args to constructor)
* simplifies default backtrace inclusion patterns
  * consistent default regardless of exclusion patterns
    * eliminates order dependency of assignments
    * does not conflict with behavior of keeping anything in working
      directory by default
* simplifies the format_backtrace helper
  * collapses it into (effectively) piped iterators
  * see benchmarks/index_v_take_while.rb